cleaning int vs. size_t in chains.

+++ addlaws.h
@@ -20,8 +20,8 @@ int pt_many_mul(ell_point_t *tQ, ell_point_t *tP, ell_curve_t *tE, int nE,
 
 int hessian_to_weierstrass(mpz_t f, mpres_t x, mpres_t y, mpres_t D, mpmod_t n);
 
-int build_MO_chain(short *S, int Slen, mpz_t e, int w);
-int build_add_sub_chain(short *S, int Slen, mpz_t e, int w);
+int build_MO_chain(short *S, size_t Slen, mpz_t e, int w);
+int build_add_sub_chain(short *S, size_t Slen, mpz_t e, int w);
 int compute_s_4_add_sub(mpz_t s, unsigned long B1, int disc);
 
 int mult_by_3(mpz_t f, mpres_t x, mpres_t y, mpres_t A, mpmod_t n);
